requirements.txt
streamlit
torch
torchvision
pillow
pandas
folium
streamlit-folium
selenium
webdriver-manager